refactor: IAsset 관련 헤더를 Types 디렉터리로 이동 및 include 경로 수정
