+++ EngineCore/Include/SimpleEngine/Asset/Loaders_DEPRECATED/IAssetLoader.h
@@ -2,7 +2,7 @@
 #include <filesystem>
 #include <memory>
 
-#include "SimpleEngine/Asset/IAsset.h"
+#include "../Types/IAsset.h"
 #include "SimpleEngine/Asset/ImportSettings_DEPRECATED/IAssetImportSettings.h"
 #include "SimpleEngine/Core/Concurrency/Coroutine.h"
 #include "SimpleEngine/Reflection/TypeSignature.h"
