+++ EngineCore/Include/SimpleEngine/Asset/Types/IAsset.h
